refactor(VtkThreeView): separate cvr/color effects
